+++ src/main/kotlin/io/unthrottled/theme/randomizer/config/Config.kt
@@ -1,21 +1,27 @@
 package io.unthrottled.theme.randomizer.config
 
+import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.PersistentStateComponent
-import com.intellij.openapi.components.ServiceManager
 import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
 import com.intellij.util.xmlb.XmlSerializerUtil.copyBean
 import com.intellij.util.xmlb.XmlSerializerUtil.createCopy
 import io.unthrottled.theme.randomizer.config.actors.LafAnimationActor
+import io.unthrottled.theme.randomizer.mode.PluginMode
+import io.unthrottled.theme.randomizer.mode.toPluginMode
 
 data class ConfigSettingsModel(
   var interval: String,
   var isChangeTheme: Boolean,
   var isRandomOrder: Boolean,
   var isThemeTransition: Boolean,
+  var pluginMode: PluginMode,
+  var changeOnSystemSwitches: Int,
 ) {
   fun duplicate(): ConfigSettingsModel = copy()
 }
+
+const val DEFAULT_OBSERVATION_COUNT = -1
 
 @State(
   name = "Theme-Randomizer-Config",
@@ -25,7 +31,7 @@ class Config : PersistentStateComponent<Config>, Cloneable {
   companion object {
     @JvmStatic
     val instance: Config
-      get() = ServiceManager.getService(Config::class.java)
+      get() = ApplicationManager.getApplication().getService(Config::class.java)
     const val DEFAULT_DELIMITER = ","
 
     @JvmStatic
@@ -34,6 +40,8 @@ class Config : PersistentStateComponent<Config>, Cloneable {
       isChangeTheme = instance.isChangeTheme,
       isRandomOrder = instance.isRandomOrder,
       isThemeTransition = LafAnimationActor.getAnimationEnabled(),
+      pluginMode = instance.pluginMode.toPluginMode(),
+      changeOnSystemSwitches = instance.changeOnSystemSwitches,
     )
   }
 
@@ -42,9 +50,17 @@ class Config : PersistentStateComponent<Config>, Cloneable {
   var isRandomOrder: Boolean = true
   var userId: String = ""
   var version: String = ""
+  var pluginMode: String = PluginMode.TIMED.displayName
   var selectedThemes = ""
   var blacklistedThemes = ""
   var lastChangeTime = -1L
+  var changeOnSystemSwitches = 1
+  var lightSystemObservedCounts = DEFAULT_OBSERVATION_COUNT
+  var darkSystemObservedCounts = DEFAULT_OBSERVATION_COUNT
+
+  fun setPluginModeEnum(pluginMode: PluginMode) {
+    this.pluginMode = pluginMode.displayName
+  }
 
   override fun getState(): Config? =
     createCopy(this)