+++ src/main/kotlin/io/unthrottled/theme/randomizer/themes/SelectableThemeType.kt
@@ -1,0 +1,5 @@
+package io.unthrottled.theme.randomizer.themes
+
+enum class SelectableThemeType {
+  DARK, LIGHT, ANY
+}