+++ docs/RELEASE-NOTES.md
@@ -1,3 +1,3 @@
 ### Added
 
-- 2022.1 Build Support!
+- Match OS mode, which allows the plugin to randomly change your theme when it observes your system settings change. [See docs for more details](./README.md#settings)